added required changes in faculty

+++ ui/faculty/faculty_delete_activity_window_logic.py
@@ -35,7 +35,12 @@ class FacultyDeleteActivityLogic(QtWidgets.QWidget):
 
     def handle_delete_activity(self):
         activity_id=self.ui.lineEdit.text()
-        response,error=self.user_dao.delete_activity(self.textbook_id,self.chapter_id,self.section_id,self.block_id,activity_id)
+        if activity_id:
+            response,error=self.user_dao.delete_activity(self.textbook_id,self.chapter_id,self.section_id,self.block_id,activity_id)
+        else:
+            QtWidgets.QMessageBox.warning(self, "Warning", 'Invalid Activity Id')
+            return
+        
         if response:
            QtWidgets.QMessageBox.information(self,'Information','Section Deleted successfully!!') 
         else:
